[CDP-12046]: OC templates support (feature branch to master) (#13031)
* [CDP-11748]: API support for OpenShift config (#12587)

* Change Log:

- Added openshift source repo type
- added validation for openshift source repo

* clang formatted

* fixed minor issue

* review feedback, minor name change

* [CDP-11748]: added support for params kind in manifest apis (#12617)

* added unit test for validation

* Change Log:

- added new params kind
- enhanced values yaml api to take kind parameter

* using default file name from enum for reducing duplication

* review feedback handling

* fixed unit test

* [CDP-11749]: Added yaml support for Params File (#12676)

* added params folder support

* changes for crud of params file through YAML

* changes for crud of params application manifest through YAML

* added params support for env. screen

* removed dead code

* clang formatted

* renaming to OC_ENTITY

* caching compiled pattern for yaml type prefix expressions

* minor bug fix, and slight code refactoring

* minor change to enummap

* clang formatted

* [CDP-11749]: added OC client layer (#12704)

* Change Log:

- added OpenShift Client and its implementation
- abstracted cli helper and response

* added non null and not empty annotations

* clang formatted

* referring to variables instead of hard-coded stuff

* minor import change

* updated tests to have cliHelper independent

* minor sonar fix - made constructor private

* redirecting errors to logger and using outputUTF8

* minor naming change

* re-arranged imports

* [CDP-11776]: added ocParams App Manifest and Manifest file to Service Template (#12874)

* [CDP-11776]: getting manifest file name irrespective of local or remote (#12905)

* [CDP-11776]: execution side changes for oc support (#12923)

* [CDP-11776]: execution side changes for oc support

* Change Log:
- Improved logging from cli
- Handling exceptions from cli

* Handled empty/null params file

* Added empty handling for oc process result and some yaml manipulation code refactoring

* renamed OpenShift Manager service and handled null application manifest scenario

* Git Service retained copy directory

* minor changes

* Change Log:
- added default file name to K8s manifest and helm kind
- Refactoring for K8s canary to use helpers
- get values/params file based on kind

* using k8shelper to get appmanifest map for all k8s states

* fixed sonar reported issues

* fixed unit tests

* fixed unit test

* added unit tests for OpenShiftClientImpl

* fixed unit for application manifest service

* [CDP-12045]: added unit tests for OpenShiftDelegateService, ManagerService, YamlType classes (#13053)

* [CDP-12045]: added unit tests for OpenShiftDelegateService

* [CDP-12045]: added unit tests for OpenShiftManagerService

* [CDP-12045]: minor fix for showing fetch files when oc template source

* [CDP-12045]: added unit test for yaml type and removed static initialization of compiled patterns

* [CDP-12045]: using computeIfAbsent as per review comments

* [CDP-12045]: added unit tests for YamlResourceService (#13098)

* [CDP-12045]: added unit tests for YamlResourceService

* [CDP-12045]: added unit tests for YamlHelper

+++ 71-rest/src/main/java/software/wings/service/impl/yaml/GitClientImpl.java
@@ -801,7 +801,11 @@ public class GitClientImpl implements GitClient {
 
         for (String filePath : gitRequest.getFilePaths()) {
           File sourceDir = new File(Paths.get(repoPath + "/" + filePath).toString());
-          FileUtils.copyDirectory(sourceDir, destinationDir);
+          if (sourceDir.isFile()) {
+            FileUtils.copyFile(sourceDir, Paths.get(destinationDirectory, filePath).toFile());
+          } else {
+            FileUtils.copyDirectory(sourceDir, destinationDir);
+          }
         }
 
         resetWorkingDir(gitConfig, gitRequest.getGitConnectorId());
@@ -816,7 +820,7 @@ public class GitClientImpl implements GitClient {
                 .append(", FilePaths: ")
                 .append(gitRequest.getFilePaths())
                 .toString(),
-            USER);
+            e, USER);
       }
     }
   }
